+++ core/src/builtin_widgets/tooltips.rs
@@ -81,11 +81,11 @@ impl<'c> ComposeChild<'c> for Tooltips {
       ));
 
       let wnd = BuildCtx::get().window();
-      let u = watch!($child.is_hover())
+      let u = watch!($child.is_hovered())
         .delay(Duration::from_millis(50), AppCtx::scheduler())
         .distinct_until_changed()
         .subscribe(move |_| {
-          if $child.is_hover() {
+          if $child.is_hovered() {
             $this.show(wnd.clone());
           } else {
             $this.hidden();